+++ minet/extraction.py
@@ -1,22 +1,21 @@
-from typing import Optional, List, Dict, Any
+from typing import Optional, List, cast
 
 from dataclasses import dataclass, field
 from casanova import TabularRecord
+from trafilatura.settings import Document as TrafilaturaDocument
 from trafilatura.core import bare_extraction
 
 from minet.exceptions import TrafilaturaError
 from minet.encodings import fix_surrogates
 
 
-def normalize_plural_trafilatura_item(meta: Dict[str, Any], key: str) -> List[str]:
-    l = meta.get(key, []) or []
-
+def normalize_plural_trafilatura_items(raw_items: Optional[List[str]]) -> List[str]:
     items = []
 
-    if not l:
+    if not raw_items:
         return items
 
-    for item in l:
+    for item in raw_items:
         if isinstance(item, dict):
             item = item.get("name")
 
@@ -45,6 +44,8 @@ def normalize_plural_trafilatura_item(meta: Dict[str, Any], key: str) -> List[st
 
 @dataclass
 class TrafilaturaResult(TabularRecord):
+    _serializer_options = {"plural_separator": "§"}
+
     canonical_url: Optional[str] = None
     title: Optional[str] = None
     description: Optional[str] = None
@@ -55,20 +56,24 @@ class TrafilaturaResult(TabularRecord):
     tags: List[str] = field(default_factory=lambda: [])
     date: Optional[str] = None
     sitename: Optional[str] = None
+    image: Optional[str] = None
+    pagetype: Optional[str] = None
 
     @classmethod
-    def from_bare_extraction(cls, data: Dict) -> "TrafilaturaResult":
+    def from_bare_extraction(cls, data: TrafilaturaDocument) -> "TrafilaturaResult":
         return TrafilaturaResult(
-            canonical_url=data.get("url"),
-            title=data.get("title"),
-            description=data.get("description"),
-            content=data.get("text"),
-            comments=data.get("comments"),
-            author=data.get("author"),
-            categories=normalize_plural_trafilatura_item(data, "categories"),
-            tags=normalize_plural_trafilatura_item(data, "tags"),
-            date=data.get("date"),
-            sitename=data.get("sitename"),
+            canonical_url=data.url,
+            title=data.title,
+            description=data.description,
+            content=data.text,
+            comments=data.comments,
+            author=data.author,
+            categories=normalize_plural_trafilatura_items(data.categories),
+            tags=normalize_plural_trafilatura_items(data.tags),
+            date=data.date,
+            sitename=data.sitename,
+            image=data.image,
+            pagetype=data.pagetype,
         )
 
     def blurb(self, fields=None) -> str:
@@ -94,5 +99,7 @@ def extract(text: str) -> Optional[TrafilaturaResult]:
 
     if trafilatura_bare_result is None:
         return
+
+    trafilatura_bare_result = cast(TrafilaturaDocument, trafilatura_bare_result)
 
     return TrafilaturaResult.from_bare_extraction(trafilatura_bare_result)