+++ CHANGELOG.md
@@ -5,10 +5,12 @@
 *Breaking*
 
 - Dropping `facebook` mobile scraping commands.
+- Upgrading `trafilatura` to v2.0.0.
 
 *Features*
 
 - Adding `reddit` commands.
+- Adding `image` and `pagetype` features to `minet extract` outputs.
 
 *Fixes*